Fix tests, process first review comments
