+++ tools/src/test/java/com/nedap/archie/rules/evaluation/ParsedRulesEvaluationTest.java
@@ -2,7 +2,6 @@ package com.nedap.archie.rules.evaluation;
 
 import com.nedap.archie.adlparser.ADLParseException;
 import com.nedap.archie.adlparser.ADLParser;
-import com.nedap.archie.adlparser.modelconstraints.RMConstraintImposer;
 import com.nedap.archie.aom.Archetype;
 import com.nedap.archie.aom.OperationalTemplate;
 import com.nedap.archie.creation.ExampleJsonInstanceGenerator;
@@ -17,38 +16,29 @@ import com.nedap.archie.rm.datastructures.Element;
 import com.nedap.archie.rm.datatypes.CodePhrase;
 import com.nedap.archie.rm.datavalues.DvCodedText;
 import com.nedap.archie.rm.datavalues.quantity.DvQuantity;
-import com.nedap.archie.rm.datavalues.quantity.datetime.DvDate;
-import com.nedap.archie.rm.datavalues.quantity.datetime.DvDateTime;
 import com.nedap.archie.rm.support.identification.TerminologyId;
 import com.nedap.archie.rminfo.ArchieRMInfoLookup;
-import com.nedap.archie.rules.Assertion;
 import com.nedap.archie.rules.BinaryOperator;
 import com.nedap.archie.rules.ExpressionVariable;
 import com.nedap.archie.rules.RuleStatement;
 import com.nedap.archie.rules.VariableDeclaration;
 import com.nedap.archie.testutil.TestUtil;
-import com.nedap.archie.xml.JAXBUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.openehr.referencemodels.BuiltinReferenceModels;
 
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 /**
  * Created by pieter.bos on 01/04/16.
  */
-public class ParsedRulesEvaluationTest {
+public abstract class ParsedRulesEvaluationTest {
 
     ADLParser parser;
     Archetype archetype;
@@ -693,80 +683,6 @@ public class ParsedRulesEvaluationTest {
         assertEquals("ac3", assertionResult.getPathsConstrainedToValueSets().get("/items[id2, 1]/items[id2]/value/defining_code"));
 
 
-    }
-
-    @Test
-    public void dateMagnitude() throws Exception {
-        parse("date_magnitude.adls");
-        RuleEvaluation<Pathable> ruleEvaluation = getRuleEvaluation();
-        Cluster root = new Cluster();
-        root.setArchetypeNodeId("id1");
-        Element element = new Element();
-        element.setArchetypeNodeId("id2");
-        element.setValue(new DvDate(LocalDate.of(2021, 6, 1)));
-        root.addItem(element);
-        ruleEvaluation.evaluate(root, archetype.getRules().getRules());
-        VariableMap variables = ruleEvaluation.getVariableMap();
-
-        assertEquals(18779l+ DvDate.DAYS_BETWEEN_0001_AND_1970, variables.get("date_magnitude").getObject(0));
-    }
-
-    @Test
-    public void dateMagnitudeDifference() throws Exception {
-        parse("date_magnitude.adls");
-        RuleEvaluation<Pathable> ruleEvaluation = getRuleEvaluation();
-        Cluster root = new Cluster();
-        root.setArchetypeNodeId("id1");
-        Element element = new Element();
-        element.setArchetypeNodeId("id2");
-        element.setValue(new DvDate(LocalDate.of(2021, 6, 20)));
-        root.addItem(element);
-        Element element2 = new Element();
-        element2.setArchetypeNodeId("id10");
-        element2.setValue(new DvDate(LocalDate.of(2021, 6, 1)));
-
-        root.addItem(element2);
-        ruleEvaluation.evaluate(root, archetype.getRules().getRules());
-        VariableMap variables = ruleEvaluation.getVariableMap();
-
-        assertEquals(19l, variables.get("date_difference").getObject(0));
-    }
-
-    @Test
-    public void dateTimeMagnitudeDifference() throws Exception {
-        parse("date_magnitude.adls");
-        RuleEvaluation<Pathable> ruleEvaluation = getRuleEvaluation();
-        Cluster root = new Cluster();
-        root.setArchetypeNodeId("id1");
-        Element element = new Element();
-        element.setArchetypeNodeId("id2");
-        element.setValue(new DvDateTime(LocalDateTime.of(2021, 6, 20, 0, 0, 40)));
-        root.addItem(element);
-        Element element2 = new Element();
-        element2.setArchetypeNodeId("id10");
-        element2.setValue(new DvDateTime(LocalDateTime.of(2021, 6, 20, 0, 0, 1)));
-
-        root.addItem(element2);
-        ruleEvaluation.evaluate(root, archetype.getRules().getRules());
-        VariableMap variables = ruleEvaluation.getVariableMap();
-
-        assertEquals(39l, variables.get("date_difference").getObject(0));
-    }
-
-    @Test
-    public void dateTimeMagnitude() throws Exception {
-        parse("date_magnitude.adls");
-        RuleEvaluation<Pathable> ruleEvaluation = getRuleEvaluation();
-        Cluster root = new Cluster();
-        root.setArchetypeNodeId("id1");
-        Element element = new Element();
-        element.setArchetypeNodeId("id2");
-        element.setValue(new DvDateTime(LocalDateTime.of(2021, 6, 1, 0, 0, 0)));
-        root.addItem(element);
-        ruleEvaluation.evaluate(root, archetype.getRules().getRules());
-        VariableMap variables = ruleEvaluation.getVariableMap();
-
-        assertEquals(63758102400l, variables.get("date_magnitude").getObject(0));
     }
 
 }
